requirements.txt
# Core
pandas>=2.0
yfinance>=0.2
requests>=2.28

# EUR/USD historical data (Phase 5)
dukascopy-python>=4.0

# Macro (Phase 6/7)
fredapi>=0.5
python-dotenv>=1.0

# News (Phase 8) — GKG raw files via requests (no gdeltdoc)

# Features (Phase 10)
transformers>=4.41
torch>=2.0
sentencepiece>=0.1.99

# Phase 11 — Training
xgboost>=2.0
scikit-learn>=1.3
# timesfm>=1.0  # Optional: requires Python <3.12; Chronos-Bolt is primary
chronos-forecasting>=2.0

# Phase 15 — Execution (demo only)
oandapyV20>=0.7
python-binance>=1.0

# Dashboard
flask>=3.0
